+++ cuvis_ai_schemas/training/__init__.py
@@ -9,7 +9,7 @@ from cuvis_ai_schemas.training.callbacks import (
     create_callbacks_from_config,
 )
 from cuvis_ai_schemas.training.config import TrainingConfig
-from cuvis_ai_schemas.training.data import DataConfig
+from cuvis_ai_schemas.training.data import DataConfig, DataSplitConfig
 from cuvis_ai_schemas.training.optimizer import OptimizerConfig
 from cuvis_ai_schemas.training.run import TrainRunConfig
 from cuvis_ai_schemas.training.scheduler import SchedulerConfig
@@ -26,6 +26,7 @@ __all__ = [
     "TrainingConfig",
     # Data
     "DataConfig",
+    "DataSplitConfig",
     # Optimizer
     "OptimizerConfig",
     # Scheduler